Extract initWindows() into its own utility file; import it into files where function is used and replace old code
