+++ packages/manager/src/features/Databases/CreateDatabaseDialog/CreateDatabaseDialog.tsx
@@ -5,8 +5,6 @@ import {
 } from '@linode/api-v4/lib/databases/types';
 import { APIError } from '@linode/api-v4/lib/types';
 import { useFormik } from 'formik';
-import { DateTime } from 'luxon';
-import { sortBy } from 'ramda';
 import * as React from 'react';
 import { useHistory } from 'react-router-dom';
 import Button from 'src/components/Button';
@@ -30,7 +28,7 @@ import { useDatabaseTypes } from 'src/hooks/useDatabaseTypes';
 import { useReduxLoad } from 'src/hooks/useReduxLoad';
 import useRegions from 'src/hooks/useRegions';
 import { getErrorStringOrDefault } from 'src/utilities/errorUtils';
-import { evenizeNumber } from 'src/utilities/evenizeNumber';
+import { initWindows } from 'src/utilities/initWindows';
 import {
   handleFieldErrors,
   handleGeneralErrors
@@ -125,23 +123,6 @@ export const CreateDatabaseDialog: React.FC<{}> = _ => {
   };
 
   // Maintenance Window
-  const initWindows = (timezone: string) => {
-    let windows = [0, 2, 4, 6, 8, 10, 12, 14, 16, 18, 20, 22].map(hour => {
-      const start = DateTime.fromObject({ hour, zone: 'utc' }).setZone(
-        timezone
-      );
-      const finish = start.plus({ hours: 2 });
-      return [
-        `${start.toFormat('HH:mm')} - ${finish.toFormat('HH:mm')}`,
-        `W${evenizeNumber(start.setZone('utc').hour)}`
-      ];
-    });
-
-    windows = sortBy<string[]>(window => window[0], windows);
-
-    return windows;
-  };
-
   const maintenanceWindowSelectOptions = React.useMemo(
     () => initWindows(timezone),
     [timezone]
